--- src/components/index.ts
@@ -1,17 +1,0 @@
-import QuickForm from "./QuickForm/QuickForm_View";
-import Loader from "./Loader";
-import Button from "./Button/Button_View";
-import Input from "./Input/Input_View";
-import QuickFormInput from "./QuickFormInput/QuickFormInput_Logic";
-import InputLabelWrapper from "./InputLabelWrapper/InputLabelWrapper_View";
-import PageLayout from "./PageLayout/PageLayout_Logic";
-
-export {
-  QuickForm,
-  Loader,
-  Button,
-  Input,
-  QuickFormInput,
-  InputLabelWrapper,
-  PageLayout
-};
