+++ src/pages/Bets.tsx
@@ -1,19 +1,19 @@
 import { useEffect, useState } from "react";
 import { useAccount } from "wagmi";
 import { useSubgraphBets } from "../hooks/subgraph";
-import { BetHistory, FilterOptions } from "../types";
-import { PageLayout } from "../components";
 import Toggle from "../components/Toggle";
 import { BetFilterGroup, BetTable } from "../components/Bets";
 import { SettleBetModal } from "../components/Modals";
-import { useWalletModal } from "src/providers/WalletModal";
+import { BetFilterOptions, BetHistory } from "../types/bets";
+import { useWalletModal } from "../providers/WalletModal";
+import PageLayout from "../components/PageLayout/PageLayout_Logic";
 
 const Bets: React.FC = () => {
   const [myBetsEnabled, setMyBetsEnabled] = useState(true);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const [selectedBet, setSelectedBet] = useState<BetHistory>();
   const [betTableFilter, setBetTableFilter] =
-    useState<FilterOptions>("ALL_BETS");
+    useState<BetFilterOptions>("ALL_BETS");
 
   const { isConnected } = useAccount();
   const { openWalletModal } = useWalletModal();
@@ -35,7 +35,8 @@ const Bets: React.FC = () => {
   }, [isConnected]);
 
   const onMyBetToggle = () => setMyBetsEnabled(prev => !prev);
-  const onFilterChange = (option: FilterOptions) => setBetTableFilter(option);
+  const onFilterChange = (option: BetFilterOptions) =>
+    setBetTableFilter(option);
 
   const isLoading = !betHistory;
   return (
