+++ src/pages/Dashboard/Dashboard_View.tsx
@@ -1,11 +1,11 @@
-import { PageLayout } from "../../components";
-import { Meet } from "../../types/index";
 import { Link } from "react-router-dom";
 import moment from "moment";
 import Skeleton from "react-loading-skeleton";
 import Toggle from "../../components/Toggle";
 import MyStats from "./components/MyStats";
 import OverallStats from "./components/OverallStats";
+import { Meet } from "../../types/meets";
+import PageLayout from "../../components/PageLayout/PageLayout_Logic";
 
 type Props = {
   asLocaltime: (raceTime: number) => string;
