+++ src/pages/HLToken/HLToken_View.tsx
@@ -1,7 +1,7 @@
-import { PageLayout } from "../../components";
 import { AiOutlineCopy } from "react-icons/ai";
 import { TokenomicsChart } from "../../images";
 import QRCode from "react-qr-code";
+import PageLayout from "../../components/PageLayout/PageLayout_Logic";
 
 const ADDRESS = "0xAA99c40A936A2cF11Eb1D5387897ce9A626E53c7";
