--- src/types/formik.ts
@@ -1,6 +1,0 @@
-import { FormikHelpers } from "formik";
-
-export type FormikOnSubmit<Values> = (
-  values: Values,
-  formikHelpers: FormikHelpers<Values>
-) => void | Promise<void>;
