+++ src/types/markets.ts
@@ -1,0 +1,7 @@
+export type Market = {
+  address: string;
+  vaultAddress: string;
+  name: string;
+  target: number;
+  totalInPlay: string;
+};
