+++ src/types/tokens.ts
@@ -1,0 +1,5 @@
+export type Token = {
+  address: string;
+  symbol: string;
+  decimals: string;
+};
