+++ src/types/users.ts
@@ -1,0 +1,7 @@
+import { BigNumber } from "ethers";
+
+export type UserBalance = {
+  value: BigNumber;
+  decimals: number;
+  formatted: string;
+};
