+++ src/types/vaults.ts
@@ -1,0 +1,25 @@
+import { VaultInfo } from "./config";
+
+export type Vault = {
+  name: string;
+  symbol: string;
+  totalAssets: string;
+  address: string;
+};
+
+export type VaultUserData = {
+  vaultBalance: string;
+  userBalance: string;
+  performance: string;
+  asset: string;
+};
+
+export enum VaultTransactionType {
+  DEPOSIT = "deposit",
+  WITHDRAW = "withdraw"
+}
+
+export type VaultModalState = {
+  type: VaultTransactionType;
+  vault: VaultInfo;
+};
